requirements.txt
requests
redis
pymysql
django
django_apscheduler
djangorestframework
